feat(ZNTA-2070) WIP add SettingsOptions and SettingOption stories to editor
