+++ server/zanata-frontend/src/frontend/app/editor/components/SettingOption/SettingOption.story.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { storiesOf, action } from '@kadira/storybook'
+import SettingOption from '.'
+
+storiesOf('SettingOption', module)
+  .add('default', () => (
+    <SettingOption
+      setting={{
+        
+      }} />
+  ))
