+++ server/zanata-frontend/src/frontend/app/editor/components/SettingsOptions/index.js
@@ -1,10 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Checkbox } from 'react-bootstrap'
+import SettingOption from '../SettingOption'
 
 const settings =
-    ['Enter key saves immediately',
-      'Syntax highlighting']
 
 const SettingsOptions = ({states, updateSettingsOption}) => {
   const checkboxes = settings.map((setting, index) => (
@@ -26,36 +25,13 @@ const SettingsOptions = ({states, updateSettingsOption}) => {
 }
 
 SettingsOptions.propTypes = {
-  states: PropTypes.shape({
-    'Enter key saves immediately': PropTypes.bool.isRequired,
-    'Syntax highlighting': PropTypes.bool.isRequired,
-    'Suggestions diff': PropTypes.bool.isRequired,
-    'Panel layout': PropTypes.bool.isRequired,
-  }).isRequired,
-  updateSettingsOption: PropTypes.func.isRequired
-}
-
-class SettingOption extends React.Component {
-  static propTypes = {
-    setting: PropTypes.string.isRequired,
-    checked: PropTypes.bool.isRequired,
-    /* Will be called with (validation, newValue) */
-    onChange: PropTypes.func.isRequired
-  }
-
-  onChange = (event) => {
-    this.props.onChange(this.props.setting, event.target.checked)
-  }
-
-  render () {
-    const { setting, checked } = this.props
-    return (
-        <Checkbox checked={checked}
-           onChange={this.onChange}>
-          {setting}
-        </Checkbox>
-    )
-  }
+  settings: PropTypes.arrayOf(shape(
+      id: PropTypes.any.isRequired, // I will update this to whatever I use when I wire it up
+      label: PropTypes.string.isRequired,
+      active: PropTypes.bool.isRequired
+  )).isRequired
+  /* arguments: (any: settingId, bool: active) */
+  updateSetting: PropTypes.func.isRequired
 }
 
 export default SettingsOptions
