+++ server/zanata-frontend/src/frontend/app/editor/components/components.story.js
@@ -11,5 +11,6 @@ require('./ProgressBar/ProgressBar.story.js')
 require('./GlossarySearchInput/GlossarySearchInput.story.js')
 require('./GlossaryTerm/GlossaryTerm.story.js')
 require('./GlossaryTermModal/GlossaryTermModal.story.js')
+require('./SettingOption/SettingOption.story.js')
 require('./SettingsOptions/SettingsOptions.story.js')
 require('./ValidationOptions/ValidationOptions.story')
